Ok I give up on doing this without numpy
FESTWERTEBLOCK blockParameter2D 4 @ 2 was the last straw. How are you supposed to keep track of dimensions without the st/y and handle if they're split because linelength limits?

Going with numpy will deal with all the annoying error handling and once the values are all parsed in a big 1d array then do a simple reshape.